🚚 Move: 설문조사 페이지 이동
