+++ src/router.tsx
@@ -6,8 +6,8 @@ import StatisticsLayout from '@/layout/StatisticsLayout';
 import ErrorPage from '@/pages/ErrorPage';
 import LandingPage from '@/pages/LandingPage';
 import RegisterPage from '@/pages/RegisterPage';
-import SurveyIntroPage from '@/pages/SurveyIntroPage';
-import SurveyPage from '@/pages/SurveyPage';
+import SurveyIntroPage from '@/pages/survey/SurveyIntroPage';
+import SurveyPage from '@/pages/survey/SurveyPage';
 import GroupCreatePage from '@/pages/group/GroupCreatePage';
 import GroupSelectPage from '@/pages/group/GroupSelectPage';
 import GroupSettingPage from '@/pages/group/GroupSettingPage';
